First Working Test Build
This is a continuation of my Large Scale Programming Project, which I originally coded in python. This is the culmination of weeks of learning SQLite (and its java implementation), and brushing up on my Java. One major change to my program as it was previously is my focus on a more object-based implementation as it makes debugging much easier (and supports what I learn in the LSP course)