Working Availability class
About 2/3 of the program is finished. Availability made me realize that it might me worth considering creating an interface to simplify the calls made in main. There are also a lot of repeated lines of code throughout the project, especially with the constant opeining of connections and preparing of SQL statements. I might implement a class or method in DbConnection to address that.